requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-dotenv>=0.19.0
PyGithub>=1.55
openai>=0.27.0
checkov>=2.3.187
pydantic>=1.8.2
python-multipart>=0.0.5
SQLAlchemy>=1.4.23
aiohttp>=3.8.1
cryptography>=3.4.8
PyYAML>=6.0
requests>=2.27.0
typing-extensions>=4.1.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.1
typer>=0.9.0
rich>=10.0.0
python-hcl2>=4.3.2
boto3>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.6.1
pytest>=7.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950
ruff>=0.1.0
pre-commit>=3.0.0 